pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codeforge"
version = "0.1.0"
description = "CodeForge — 编码能力锻造 CLI 工具"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Eric", email = "eric@example.com"},
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
]

[project.scripts]
forge = "codeforge.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/codeforge"]
